requirements.txt
python-frontmatter>=0.5.0
python-dateutil>=2.8.1
pytz==2019.1
git+git://github.com/lukasschwab/jsonfeed.git#egg=jsonfeed
